Revert "Move conversation menu options to a dialog"
