+++ src/components/RightSidebar/Participants/ParticipantsTab.vue
@@ -48,7 +48,7 @@ import CurrentParticipants from './CurrentParticipants/CurrentParticipants'
 import SearchBox from '../../LeftSidebar/SearchBox/SearchBox'
 import debounce from 'debounce'
 import { EventBus } from '../../../services/EventBus'
-import { PARTICIPANT } from '../../../constants'
+import { CONVERSATION, PARTICIPANT, WEBINAR } from '../../../constants'
 import { searchPossibleConversations } from '../../../services/conversationsService'
 import {
 	addParticipant,
@@ -119,7 +119,16 @@ export default {
 			return this.$store.getters.getToken()
 		},
 		conversation() {
-			return this.$store.getters.conversation(this.token) || this.$store.getters.dummyConversation
+			if (this.$store.getters.conversation(this.token)) {
+				return this.$store.getters.conversation(this.token)
+			}
+			return {
+				token: '',
+				displayName: '',
+				isFavorite: false,
+				type: CONVERSATION.TYPE.PUBLIC,
+				lobbyState: WEBINAR.LOBBY.NONE,
+			}
 		},
 		isSearching() {
 			return this.searchText !== ''
