+++ src/components/TopBar/TopBar.vue
@@ -81,6 +81,14 @@
 					@click="handleRenameConversation">
 					{{ t('spreed', 'Rename conversation') }}
 				</ActionButton>
+				<ActionSeparator
+					v-if="canFullModerate" />
+				<ActionCheckbox
+					v-if="canFullModerate"
+					:checked="isSharedPublicly"
+					@change="toggleGuests">
+					{{ t('spreed', 'Share link') }}
+				</ActionCheckbox>
 			</template>
 			<ActionButton
 				v-if="!isOneToOneConversation"
@@ -89,6 +97,56 @@
 				@click="handleCopyLink">
 				{{ t('spreed', 'Copy link') }}
 			</ActionButton>
+			<!-- password -->
+			<template
+				v-if="showModerationOptions">
+				<ActionCheckbox
+					v-if="isSharedPublicly"
+					class="share-link-password-checkbox"
+					:checked="isPasswordProtected"
+					@check="handlePasswordEnable"
+					@uncheck="handlePasswordDisable">
+					{{ t('spreed', 'Password protection') }}
+				</ActionCheckbox>
+				<ActionInput
+					v-show="isEditingPassword"
+					class="share-link-password"
+					icon="icon-password"
+					type="password"
+					:value.sync="password"
+					autocomplete="new-password"
+					@submit="handleSetNewPassword">
+					{{ t('spreed', 'Enter a password') }}
+				</ActionInput>
+				<ActionSeparator />
+				<ActionCheckbox
+					:checked="isReadOnly"
+					:disabled="readOnlyStateLoading"
+					@change="toggleReadOnly">
+					{{ t('spreed', 'Lock conversation') }}
+				</ActionCheckbox>
+				<ActionCheckbox
+					:checked="hasLobbyEnabled"
+					@change="toggleLobby">
+					{{ t('spreed', 'Enable lobby') }}
+				</ActionCheckbox>
+				<ActionInput
+					v-if="hasLobbyEnabled"
+					icon="icon-calendar-dark"
+					type="datetime-local"
+					v-bind="dateTimePickerAttrs"
+					:value="lobbyTimer"
+					:disabled="lobbyTimerLoading"
+					@change="setLobbyTimer">
+					{{ t('spreed', 'Start time (optional)') }}
+				</ActionInput>
+				<ActionCheckbox
+					v-if="canUserEnableSIP"
+					:checked="hasSIPEnabled"
+					@change="toggleSIPEnabled">
+					{{ t('spreed', 'Enable SIP dial-in') }}
+				</ActionCheckbox>
+			</template>
 			<template
 				v-if="showModerationOptions && canFullModerate && isInCall">
 				<ActionSeparator />
@@ -99,15 +157,6 @@
 					{{ t('spreed', 'Mute others') }}
 				</ActionButton>
 			</template>
-			<ActionSeparator
-				v-if="showModerationOptions" />
-			<ActionButton
-				v-if="showModerationOptions"
-				icon="icon-settings"
-				:close-after-click="true"
-				@click="showConversationSettings">
-				{{ t('spreed', 'More settings') }}
-			</ActionButton>
 		</Actions>
 		<Actions v-if="showOpenSidebarButton"
 			class="top-bar__button"
@@ -127,12 +176,16 @@ import Actions from '@nextcloud/vue/dist/Components/Actions'
 import CallButton from './CallButton'
 import { EventBus } from '../../services/EventBus'
 import BrowserStorage from '../../services/BrowserStorage'
+import ActionCheckbox from '@nextcloud/vue/dist/Components/ActionCheckbox'
+import ActionInput from '@nextcloud/vue/dist/Components/ActionInput'
 import ActionLink from '@nextcloud/vue/dist/Components/ActionLink'
 import ActionSeparator from '@nextcloud/vue/dist/Components/ActionSeparator'
-import { CONVERSATION, PARTICIPANT } from '../../constants'
+import { CONVERSATION, WEBINAR, PARTICIPANT } from '../../constants'
+import {
+	setConversationPassword,
+} from '../../services/conversationsService'
 import { generateUrl } from '@nextcloud/router'
 import { callParticipantCollection } from '../../utils/webrtc/index'
-import { emit } from '@nextcloud/event-bus'
 
 export default {
 	name: 'TopBar',
@@ -140,6 +193,8 @@ export default {
 	components: {
 		ActionButton,
 		Actions,
+		ActionCheckbox,
+		ActionInput,
 		ActionLink,
 		CallButton,
 		Popover,
@@ -157,6 +212,12 @@ export default {
 		return {
 			showLayoutHint: false,
 			hintDismissed: false,
+			// The conversation's password
+			password: '',
+			// Switch for the password-editing operation
+			isEditingPassword: false,
+			lobbyTimerLoading: false,
+			readOnlyStateLoading: false,
 		}
 	},
 
@@ -236,11 +297,29 @@ export default {
 		showModerationOptions() {
 			return !this.isOneToOneConversation && this.canModerate
 		},
+		canUserEnableSIP() {
+			return this.conversation.canEnableSIP
+		},
 		token() {
 			return this.$store.getters.getToken()
 		},
+		isSharedPublicly() {
+			return this.conversation.type === CONVERSATION.TYPE.PUBLIC
+		},
 		conversation() {
-			return this.$store.getters.conversation(this.token) || this.$store.getters.dummyConversation
+			if (this.$store.getters.conversation(this.token)) {
+				return this.$store.getters.conversation(this.token)
+			}
+			return {
+				token: '',
+				displayName: '',
+				isFavorite: false,
+				hasPassword: false,
+				canEnableSIP: false,
+				type: CONVERSATION.TYPE.PUBLIC,
+				lobbyState: WEBINAR.LOBBY.NONE,
+				lobbyTimer: 0,
+			}
 		},
 		linkToConversation() {
 			if (this.token !== '') {
@@ -253,7 +332,47 @@ export default {
 			return this.conversation.type === CONVERSATION.TYPE.GROUP
 				|| this.conversation.type === CONVERSATION.TYPE.PUBLIC
 		},
+		hasLobbyEnabled() {
+			return this.conversation.lobbyState === WEBINAR.LOBBY.NON_MODERATORS
+		},
+		isReadOnly() {
+			return this.conversation.readOnly === CONVERSATION.STATE.READ_ONLY
+		},
+		isPasswordProtected() {
+			return this.conversation.hasPassword
+		},
+		lobbyTimer() {
+			// A timestamp of 0 means that there is no lobby, but it would be
+			// interpreted as the Unix epoch by the DateTimePicker.
+			if (this.conversation.lobbyTimer === 0) {
+				return undefined
+			}
 
+			// PHP timestamp is second-based; JavaScript timestamp is
+			// millisecond based.
+			return new Date(this.conversation.lobbyTimer * 1000)
+		},
+
+		dateTimePickerAttrs() {
+			return {
+				format: 'YYYY-MM-DD HH:mm',
+				firstDayOfWeek: window.firstDay + 1, // Provided by server
+				lang: {
+					days: window.dayNamesShort, // Provided by server
+					months: window.monthNamesShort, // Provided by server
+				},
+				// Do not update the value until the confirm button has been
+				// pressed. Otherwise it would not be possible to set a lobby
+				// for today, because as soon as the day is selected the lobby
+				// timer would be set, but as no time was set at that point the
+				// lobby timer would be set to today at 00:00, which would
+				// disable the lobby due to being in the past.
+				confirm: true,
+			}
+		},
+		hasSIPEnabled() {
+			return this.conversation.sipEnabled === WEBINAR.SIP.ENABLED
+		},
 		isGrid() {
 			return this.$store.getters.isGrid
 		},
@@ -281,10 +400,6 @@ export default {
 		openSidebar() {
 			this.$store.dispatch('showSidebar')
 			BrowserStorage.setItem('sidebarOpen', 'true')
-		},
-
-		showConversationSettings() {
-			emit('show-conversation-settings')
 		},
 
 		fullScreenChanged() {
@@ -335,7 +450,77 @@ export default {
 			this.$store.dispatch('selectedVideoPeerId', null)
 			this.showLayoutHint = false
 		},
+		async toggleGuests() {
+			await this.$store.dispatch('toggleGuests', {
+				token: this.token,
+				allowGuests: this.conversation.type !== CONVERSATION.TYPE.PUBLIC,
+			})
+		},
 
+		async toggleLobby() {
+			await this.$store.dispatch('toggleLobby', {
+				token: this.token,
+				enableLobby: this.conversation.lobbyState !== WEBINAR.LOBBY.NON_MODERATORS,
+			})
+		},
+
+		async toggleSIPEnabled(checked) {
+			try {
+				await this.$store.dispatch('setSIPEnabled', {
+					token: this.token,
+					state: checked ? WEBINAR.SIP.ENABLED : WEBINAR.SIP.DISABLED,
+				})
+			} catch (e) {
+				// TODO check "precondition failed"
+				// TODO showError()
+				console.error(e)
+			}
+		},
+
+		async setLobbyTimer(date) {
+			this.lobbyTimerLoading = true
+
+			let timestamp = 0
+			if (date) {
+				// PHP timestamp is second-based; JavaScript timestamp is
+				// millisecond based.
+				timestamp = date.getTime() / 1000
+			}
+
+			await this.$store.dispatch('setLobbyTimer', {
+				token: this.token,
+				timestamp: timestamp,
+			})
+
+			this.lobbyTimerLoading = false
+		},
+
+		async toggleReadOnly() {
+			this.readOnlyStateLoading = true
+			await this.$store.dispatch('setReadOnlyState', {
+				token: this.token,
+				readOnly: this.isReadOnly ? CONVERSATION.STATE.READ_WRITE : CONVERSATION.STATE.READ_ONLY,
+			})
+			this.readOnlyStateLoading = false
+		},
+
+		async handlePasswordDisable() {
+			// disable the password protection for the current conversation
+			if (this.conversation.hasPassword) {
+				await setConversationPassword(this.token, '')
+			}
+			this.password = ''
+			this.isEditingPassword = false
+		},
+		async handlePasswordEnable() {
+			this.isEditingPassword = true
+		},
+
+		async handleSetNewPassword() {
+			await setConversationPassword(this.token, this.password)
+			this.password = ''
+			this.isEditingPassword = false
+		},
 		async handleCopyLink() {
 			try {
 				await this.$copyText(this.linkToConversation)
